Identity contract: Move types to a separate file
